Extract useImagesInfinite types into types.ts
Pure mechanical move of type/interface definitions from useImagesInfinite.ts
to a new types.ts file. No logic changes. Sets up further refactors that
can now reference types independently of the hook implementation.
